+++ Source_Files/Misc/sdl_dialogs.h
@@ -263,7 +263,7 @@ extern void load_theme(FileSpecifier &theme);
 
 extern const sdl_font_info *get_dialog_font(int which, uint16 &style);
 extern uint32 get_dialog_color(int which);
-extern uint32 get_dialog_player_color(int colorIndex); // ZZZ: added
+extern uint32 get_dialog_player_color(size_t colorIndex); // ZZZ: added
 extern SDL_Surface *get_dialog_image(int which, int width = 0, int height = 0);
 extern int get_dialog_space(int which);
 extern void play_dialog_sound(int which);
